+++ demo3.php
@@ -1,8 +1,8 @@
 <?php
 
 use GuzzleHttp\Client;
-use paslandau\GuzzleAutoCharsetEncodingSubscriber\EncodingConverter;
 use paslandau\GuzzleAutoCharsetEncodingSubscriber\GuzzleAutoCharsetEncodingSubscriber;
+use paslandau\WebUtility\EncodingConversion\EncodingConverter;
 
 require_once __DIR__ . '/demo-bootstrap.php';